requirements.txt
fastapi>=0.112.0,<0.113.0
uvicorn[standard]>=0.30.0,<0.31.0
sqlalchemy>=2.0.32,<2.1.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
python-multipart>=0.0.9,<0.1.0
pillow>=10.4.0,<11.0.0
numpy>=1.26.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
loguru>=0.7.2,<0.8.0
pytest>=8.3.0,<9.0.0
httpx>=0.27.0,<0.28.0
torch>=2.7.0,<3.0.0
torchvision>=0.22.0,<0.23.0
torchaudio>=2.7.0,<3.0.0
ultralytics>=8.3.0,<9.0.0
typing-extensions>=4.10.0,<5.0.0
